refactor: post-overlay-refactor simplify pass + 0.95.0 version bump
- Introduce OverlayMode string union + OVERLAY_MODE_TO_KIND lookup; type
  _overlayMode fields in panel and sidebar
- Collapse overlay stripe CSS into OVERLAY_STRIPE_CSS lookup
- Extract hasEntryOverlayNear helper; replace two duplicated 3x3 scans in
  zone-engine.ts
- Hoist shared mouseup listener setup in onCellMouseDown; add no-op
  guard to applyPaintToCell to skip copy+re-render when cell unchanged
- Firmware: overlay kind constants → uint8_t for consistency with mask;
  reuse existing overlay local instead of second cell_overlay() call
- Bump firmware + integration to 0.95.0 (breaking storage format change)

+++ firmware/lib/epp_zone_engine/include/epp_types.h
@@ -19,10 +19,10 @@ constexpr uint8_t CELL_ZONE_MASK = 0x0E;
 // Overlay field — bits 4-5, kind 0..3
 constexpr uint8_t CELL_OVERLAY_MASK = 0x30;
 constexpr int CELL_OVERLAY_SHIFT = 4;
-constexpr int CELL_OVERLAY_NONE = 0;
-constexpr int CELL_OVERLAY_ENTRY = 1;
-constexpr int CELL_OVERLAY_INTERFERENCE = 2;
-constexpr int CELL_OVERLAY_SUPPRESS = 3;
+constexpr uint8_t CELL_OVERLAY_NONE = 0;
+constexpr uint8_t CELL_OVERLAY_ENTRY = 1;
+constexpr uint8_t CELL_OVERLAY_INTERFERENCE = 2;
+constexpr uint8_t CELL_OVERLAY_SUPPRESS = 3;
 
 // Limits
 constexpr int MAX_TARGETS = 3;
